+++ wgpu-hal/src/gles/adapter.rs
@@ -792,6 +792,10 @@ impl super::Adapter {
                 }
             },
             max_color_attachments,
+            max_subpass_color_attachments: 0,
+            max_subpasses: 0,
+            max_input_attachments: 0,
+            estimated_tile_memory_bytes: 0,
             max_color_attachment_bytes_per_sample,
             max_compute_workgroup_storage_size: if supports_work_group_params {
                 (unsafe { gl.get_parameter_i32(glow::MAX_COMPUTE_SHARED_MEMORY_SIZE) } as u32)